pyproject.toml
[project]
name = "clang-tidy-wrap"
version = "0.1.0"
description = "AI-friendly clang-tidy wrapper for compile database driven C/C++ analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "orjson>=3.10",
  "pydantic>=2.8",
  "PyYAML>=6.0",
  "typer>=0.12",
]

[project.scripts]
ctwrap = "ctwrap.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ctwrap"]

[tool.pytest.ini_options]
testpaths = ["tests"]
